test: Switching dropoff location to test the code
The drop off location for the second marker was changed to test the code. The marker location and the goal location must be specified. Therefore, changes to the location of the marker must be simultaneously changed with the pick_objects.cpp. The locations of the markers must also be constrained to the world map.